move table/notice-table
