+++ packages/insomnia/src/ui/insomnia-components/index.ts
@@ -6,18 +6,6 @@ export { CardContainer } from './card-container';
 export { Card, type  CardProps } from './card';
 export { Header, type  HeaderProps } from './header';
 export { MultiSwitch } from './multi-switch';
-export { NoticeTable, type NoticeTableProps, type Notice } from './notice-table';
 export { RadioButtonGroup, type RadioButtonGroupProps } from './radio-button-group';
 export { SvgIcon, type SvgIconProps, type IconEnum, type ThemeEnum, type IconId } from './svg-icon';
 export { Switch, type SwitchProps, type SwitchItem } from './switch';
-export {
-  Table,
-  type TableProps,
-  TableBody,
-  TableData,
-  type TableDataProps,
-  TableHead,
-  TableHeader,
-  type TableHeaderProps,
-  TableRow,
-} from './table';
